test: Add uncertainity to the code for testing purposes
In my opinion, the configuration object should always have the options
key and it's responsibility of the reporter class or user of the class
to ensure that. However, current code is written in a different way
and forces the reporter to use the @config property in a very deffensive
way. Since I don't want to rewrite the whole thing, I'm following the
practice with this commit, adding uncertainity to the code.